# swagger-bundle.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FilmReview API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password",
                  "fullName",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "400": {
            "description": "Erro na validação"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/auth/{id}": {
      "delete": {
        "summary": "Excluir usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário removido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      },
      "put": {
        "summary": "Atualizar informações do usuário (todos os campos opcionais)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "example": "Usuário da Silva"
                  },
                  "email": {
                    "type": "string",
                    "example": "usuario123@user.com"
                  },
                  "photo": {
                    "type": "string",
                    "example": "profile-123.jpg"
                  },
                  "pronouns": {
                    "type": "string",
                    "example": "ele/dele"
                  },
                  "bio": {
                    "type": "string",
                    "example": "Sou apaixonado por cinema."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Perfil atualizado com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "msg": {
                            "type": "string",
                            "example": "Biografia deve ter até 100 caracteres"
                          },
                          "param": {
                            "type": "string",
                            "example": "bio"
                          },
                          "location": {
                            "type": "string",
                            "example": "body"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário não encontrado"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies": {
      "get": {
        "summary": "Buscar filmes pelo nome",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nome do filme para busca"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes e contagem de reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "photo": {
                        "type": "string"
                      },
                      "reviewCount": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetro name obrigatório"
          }
        }
      },
      "post": {
        "summary": "Criar novo filme",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name",
                  "photo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Filme criado com sucesso"
          },
          "400": {
            "description": "Erro na validação"
          }
        }
      }
    },
    "/reviews": {
      "get": {
        "summary": "Buscar reviews pelo nome do filme",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nome do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reviews do filme",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "username": {
                        "type": "string"
                      },
                      "rating": {
                        "type": "number",
                        "format": "float"
                      },
                      "reviewText": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetro name obrigatório"
          }
        }
      },
      "post": {
        "summary": "Criar nova review",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "movieId": {
                    "type": "integer"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "reviewText": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "movieId",
                  "rating",
                  "reviewText"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review criada com sucesso"
          },
          "400": {
            "description": "Erro na validação"
          }
        }
      }
    },
    "/reviews/{id}": {
      "delete": {
        "summary": "Excluir review",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Review removida"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Review não encontrada"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    }
  }
}